Refactor and restructure files.
