+++ benchmarks/hash_map_benchmark.cpp
@@ -1,200 +1,12 @@
 #include "benchmark.hpp"
 #include "execution.hpp"
 #include "config.hpp"
-#include "descriptor.hpp"
+#include "hash_maps.hpp"
 
 #include <iostream>
 #include <vector>
 
-#ifdef WITH_VYUKOV_HASH_MAP
-#include <xenium/vyukov_hash_map.hpp>
-#endif
-
-#ifdef WITH_HARRIS_MICHAEL_HASH_MAP
-#include <xenium/harris_michael_hash_map.hpp>
-#endif
-
-#ifdef WITH_LIBCDS
-#include <cds/gc/hp.h>
-#include <cds/gc/dhp.h>
-#endif
-
-#ifdef WITH_CDS_MICHAEL_MAP
-#include <cds/container/michael_kvlist_hp.h>
-#include <cds/container/michael_map.h>
-#endif
-
-#ifdef WITH_CDS_FELDMAN_HASHMAP
-#include <cds/container/feldman_hashmap_hp.h>
-#include <cds/container/feldman_hashmap_dhp.h>
-#include <cds/container/feldman_hashmap_rcu.h>
-#endif
-
 using boost::property_tree::ptree;
-
-template <class T>
-struct hash_map_builder {
-  static auto create(const boost::property_tree::ptree&) { return std::make_unique<T>(); }
-};
-
-template <class T>
-struct region_guard {
-  using type = typename T::reclaimer::region_guard;
-};
-
-template <class T>
-using region_guard_t = typename region_guard<T>::type;
-
-#ifdef WITH_VYUKOV_HASH_MAP
-template <class Key, class Value, class... Policies>
-struct descriptor<xenium::vyukov_hash_map<Key, Value, Policies...>> {
-  static boost::property_tree::ptree generate() {
-    using hash_map = xenium::vyukov_hash_map<Key, Value, Policies...>;
-    boost::property_tree::ptree pt;
-    pt.put("type", "vyukov_hash_map");
-    pt.put("initial_capacity", DYNAMIC_PARAM);
-    pt.put_child("reclaimer", descriptor<typename hash_map::reclaimer>::generate());
-    return pt;
-  }
-};
-
-template <class Key, class Value, class... Policies>
-struct hash_map_builder<xenium::vyukov_hash_map<Key, Value, Policies...>> {
-  static auto create(const boost::property_tree::ptree& config) {
-    auto initial_capacity = config.get<size_t>("initial_capacity", 128);
-    return std::make_unique<xenium::vyukov_hash_map<Key, Value, Policies...>>(initial_capacity);
-  }
-};
-
-namespace {
-  template <class Key, class Value, class... Policies>
-  bool try_emplace(xenium::vyukov_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    return hash_map.emplace(key, key);
-  }
-
-  template <class Key, class Value, class... Policies>
-  bool try_remove(xenium::vyukov_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    return hash_map.erase(key);
-  }
-
-  template <class Key, class Value, class... Policies>
-  bool try_get(xenium::vyukov_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    typename xenium::vyukov_hash_map<Key, Value, Policies...>::accessor acc;
-    return hash_map.try_get_value(key, acc);
-  }
-}
-#endif
-
-#ifdef WITH_HARRIS_MICHAEL_HASH_MAP
-template <class Key, class Value, class... Policies>
-struct descriptor<xenium::harris_michael_hash_map<Key, Value, Policies...>> {
-  static boost::property_tree::ptree generate() {
-    using hash_map = xenium::harris_michael_hash_map<Key, Value, Policies...>;
-    boost::property_tree::ptree pt;
-    pt.put("type", "harris_michael_hash_map");
-    auto buckets = hash_map::num_buckets;
-    pt.put("buckets", buckets);
-    pt.put_child("reclaimer", descriptor<typename hash_map::reclaimer>::generate());
-    return pt;
-  }
-};
-
-namespace {
-  template <class Key, class Value, class... Policies>
-  bool try_emplace(xenium::harris_michael_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    return hash_map.emplace(key, key);
-  }
-
-  template <class Key, class Value, class... Policies>
-  bool try_remove(xenium::harris_michael_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    return hash_map.erase(key);
-  }
-
-  template <class Key, class Value, class... Policies>
-  bool try_get(xenium::harris_michael_hash_map<Key, Value, Policies...>& hash_map, std::uint64_t key) {
-    auto it = hash_map.find(key);
-    return it != hash_map.end();
-  }
-}
-#endif
-
-
-#ifdef WITH_CDS_MICHAEL_MAP
-template <class GC, class List, class Traits>
-struct descriptor<cds::container::MichaelHashMap<GC, List, Traits>> {
-  static boost::property_tree::ptree generate() {
-    boost::property_tree::ptree pt;
-    pt.put("type", "cds::MichaelMap");
-    pt.put("nMaxItemCount", DYNAMIC_PARAM);
-    pt.put("nLoadFactor", DYNAMIC_PARAM);  
-    return pt;
-  }
-};
-
-template <class GC, class List, class Traits>
-struct hash_map_builder<cds::container::MichaelHashMap<GC, List, Traits>> {
-  static auto create(const boost::property_tree::ptree& config) {
-    auto nMaxItemCount = config.get<size_t>("nMaxItemCount");
-    auto nLoadFactor = config.get<size_t>("nLoadFactor");
-    return std::make_unique<cds::container::MichaelHashMap<GC, List, Traits>>(nMaxItemCount, nLoadFactor);
-  }
-};
-
-template <class GC, class List, class Traits>
-struct region_guard<cds::container::MichaelHashMap<GC, List, Traits>> {
-  struct type{};
-};
-
-namespace {
-  template <class GC, class List, class Traits>
-  bool try_emplace(cds::container::MichaelHashMap<GC, List, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.emplace(key, key);
-  }
-
-  template <class GC, class List, class Traits>
-  bool try_remove(cds::container::MichaelHashMap<GC, List, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.erase(key);
-  }
-
-  template <class GC, class List, class Traits>
-  bool try_get(cds::container::MichaelHashMap<GC, List, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.find(key, [](auto& /*val*/){});
-  }
-}
-#endif
-
-#ifdef WITH_CDS_FELDMAN_HASHMAP
-template <class GC, class Key, class T, class Traits>
-struct descriptor<cds::container::FeldmanHashMap<GC, Key, T, Traits>> {
-  static boost::property_tree::ptree generate() {
-    boost::property_tree::ptree pt;
-    pt.put("type", "cds::FeldmanHashMap");
-    return pt;
-  }
-};
-
-template <class GC, class Key, class T, class Traits>
-struct region_guard<cds::container::FeldmanHashMap<GC, Key, T, Traits>> {
-  struct type{};
-};
-
-namespace {
-  template <class GC, class Key, class T, class Traits>
-  bool try_emplace(cds::container::FeldmanHashMap<GC, Key, T, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.emplace(key, key);
-  }
-
-  template <class GC, class Key, class T, class Traits>
-  bool try_remove(cds::container::FeldmanHashMap<GC, Key, T, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.erase(key);
-  }
-
-  template <class GC, class Key, class T, class Traits>
-  bool try_get(cds::container::FeldmanHashMap<GC, Key, T, Traits>& hash_map, std::uint64_t key) {
-    return hash_map.find(key, [](auto& /*val*/){});
-  }
-}
-#endif
 
 template <class T>
 struct hash_map_benchmark;
